Fix api access view review feedback
